modularizes server
